**** TEST COMMIT **** DO NOT APPLY ****
We had that absolutely crazy construct, where in some QML-accessed
singletons the instance variable was set in the constructor. The
background was that apparently if registered as a normal type, QML
would insist on constructing an object via "new T".

Nevertheless, as a quick look in the documentation shows, QML *does*
support singletons with the qmlRegisterSingletonType() function.
This is a first proof of concept to make the QMLPrefs class a real
singleton.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>