Bluetooth: don't construct address string in QString on Windows
The old code passed a pointer to the data of a pre-initialized
QString to WSAAddressToStringW(). This is valid, but somewhat
uncanny, as it creates a string with a 0-character. The 0-character
will then truncated away by cutting the address to a canonical
length. Still a scary proposition.

Instead, pass a wchar-array to WSAAddressToStringW() and construct
a string from that. This makes the code cleaner as it removes
an unnecessary cast and one doesn't have to worry about 0-characters
in strings.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>