Cleanup: don't use QByteArray::data() to create copy
QByteArray::data() provides access to the underlying data
for direct manipulation. Thus, the construct
	csv = fileNamePtr.data();
found in MainWindow::importTxtFiles() suggests that modifications
to csv also affect fileNamePtr. This is *not* the case, because
csv itself is a QByteArray. It is therefore constructed from
the data.

Replace this treacherous construct by a simple assignment.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>