Cleanup: don't derive UserSurveyServices from WebServices
UserSurveyServices derives from WebServices and therefore has
to define three pure virtual functions [startDownload(),
startUpload(), buttonClicked()] as no-ops. Interestingly,
a comment in the header says
	"need to declare them as no ops or Qt4 is unhappy"
which is of course not true as these functions are not
declared by Qt.

There seems to be no point in deriving from WebServices,
therefore don't do it. These function definitions can then
be removed.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>